Separated R1->C into a new python file, independent of S->C conversion.
